backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
websockets==12.0
httpx==0.25.2
fastapi-cors==0.0.6
psycopg2-binary==2.9.9
gunicorn==21.2.0
